rename: 오늘 지출 추천 시 재정 상태 나타내는 FinanceStatus enum 패키지 변경
- web → business 계층으로 이동: 비즈니스 로직에 더 가까운 값이므로

[SW-30]
