+++ src/main/java/com/wanted/safewallet/domain/expenditure/business/dto/TodayExpenditureConsultDto.java
@@ -1,6 +1,6 @@
 package com.wanted.safewallet.domain.expenditure.business.dto;
 
-import com.wanted.safewallet.domain.expenditure.web.enums.FinanceStatus;
+import com.wanted.safewallet.domain.expenditure.business.enums.FinanceStatus;
 
 public record TodayExpenditureConsultDto(Long amount, FinanceStatus financeStatus) {
 }
